+++ java/code/src/com/redhat/rhn/domain/rhnpackage/profile/test/ProfileTest.java
@@ -72,8 +72,8 @@ public class ProfileTest extends RhnBaseTestCase {
     public static Profile lookupByIdAndOrg(Long id, Org org) throws Exception {
         Session session = HibernateFactory.getSession();
         return (Profile) session.getNamedQuery("Profile.findByIdAndOrg")
-                                    .setLong("id", id.longValue())
-                                    .setLong("org_id", org.getId().longValue())
+                                    .setLong("id", id)
+                                    .setLong("org_id", org.getId())
                                     .uniqueResult();
     }
 
@@ -119,8 +119,8 @@ public class ProfileTest extends RhnBaseTestCase {
         session.flush();
 
         Query qry = session.getNamedQuery("Profile.compatibleWithServer");
-        qry.setLong("sid", server.getId().longValue());
-        qry.setLong("org_id", user.getOrg().getId().longValue());
+        qry.setLong("sid", server.getId());
+        qry.setLong("org_id", user.getOrg().getId());
         List list = qry.list();
         assertNotNull("List is null", list);
         assertFalse("List is empty", list.isEmpty());
