+++ java/code/src/com/redhat/rhn/frontend/dto/ChannelTreeNode.java
@@ -363,7 +363,7 @@ public class ChannelTreeNode extends BaseDto implements BaseListDto,
      * {@inheritDoc}
      */
     public long depth() {
-        return getDepth().longValue();
+        return getDepth();
     }
