+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/test/SatScrubberTest.java
@@ -177,7 +177,7 @@ public class SatScrubberTest extends TestCase {
             Long id = (Long) row.get("id");
             log.debug("Deleting org: " + id);
             try {
-                OrgFactory.deleteOrg(id.longValue(), UserFactory
+                OrgFactory.deleteOrg(id, UserFactory
                         .findResponsibleUser(1L, RoleFactory.SAT_ADMIN));
             }
             catch (Exception e) {
